# specs/001-local-prompt-manager/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Prompt Management Local API
  version: 1.0.0
  description: 本機提示詞管理 Route Handlers 合約（Next.js REST）
servers:
  - url: http://localhost:3000/api
paths:
...
  /snippets:
    get:
      summary: 列出片語（可搜尋）
      parameters:
        - in: query
          name: q
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Snippet"
    post:
      summary: 新增片語
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SnippetCreateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Snippet"
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
    patch:
      summary: 更新片語（collection 形式；以 body.id 指定）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SnippetUpdateByBodyInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Snippet"
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      summary: 刪除片語（collection 形式；以 body.id 指定）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
...
  responses:
    BadRequest:
      description: 驗證失敗
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: 資源不存在
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Conflict:
      description: 外部修改或版本衝突
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code: { type: string, description: e.g. bad_request | not_found | conflict }
        message: { type: string }
        details: { description: 錯誤細節（依端點而異） }
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
    IsoDateTime:
      type: string
      format: date-time
      description: ISO 8601（UTC+08:00）
...
    Snippet:
      type: object
      required: [id, name, category, content, usageCount, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string }
        content: { type: string }
        usageCount: { type: integer }
        usage: { type: integer }
        lastUsedAt: { $ref: "#/components/schemas/IsoDateTime" }
        createdAt: { $ref: "#/components/schemas/IsoDateTime" }
        updatedAt: { $ref: "#/components/schemas/IsoDateTime" }
    SnippetCreateInput:
      type: object
      required: [name]
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string }
        content: { type: string }
    SnippetUpdateByBodyInput:
      type: object
      required: [id]
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string }
        content: { type: string }
...